requirements.txt
Pillow>=10.0.0
tkinterdnd2>=0.3.0
docx2pdf>=0.1.8
pillow-heif>=0.15.0
pyinstaller>=6.0.0 
psutil>=5.9.0
typing-extensions>=4.0.0 